Code cleanup involving enums and message replies
